+++ beacons/handlers.go
@@ -1,0 +1,191 @@
+// Copyright 2014 Caleb Brose, Chris Fogerty, Rob Sheehy, Zach Taylor, Nick Miller
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package beacons
+
+import (
+	"fmt"
+	"errors"
+	"net/http"
+	"encoding/json"
+	"io/ioutil"
+
+	"github.com/gorilla/mux"
+
+	beaconStructs "github.com/lighthouse/beacon/structs"
+
+    "github.com/lighthouse/lighthouse/beacons/aliases"
+    "github.com/lighthouse/lighthouse/databases"
+)
+
+func getInstanceAlias(instance string) string {
+    alias, err := aliases.GetAlias(instance)
+    if err != nil {
+        return instance
+    }
+    return alias
+}
+
+func writeResponse(err error, w http.ResponseWriter) {
+    var code int
+
+    switch err {
+        case databases.KeyNotFoundError, databases.NoUpdateError, databases.EmptyKeyError:
+            code = http.StatusBadRequest
+
+        case nil:
+            code = http.StatusOK
+
+        default:
+            code = http.StatusInternalServerError
+    }
+
+    w.WriteHeader(code)
+
+    if err != nil {
+        fmt.Fprint(w, err)
+    }
+}
+
+func handleAddUserToBeacon(w http.ResponseWriter, r *http.Request) {
+    vars := mux.Vars(r)
+
+    instance := getInstanceAlias(vars["Instance"])
+    userId := vars["Id"]
+
+    beacon, err := getBeaconData(instance)
+
+    if err == nil {
+        if beacon.Users == nil {
+            beacon.Users = userMap{userId : true}
+        } else {
+            beacon.Users[userId] = true
+        }
+        err = updateBeaconField("Users", beacon.Users, instance)
+    }
+
+    writeResponse(err, w)
+}
+
+func handleRemoveUserFromBeacon(w http.ResponseWriter, r *http.Request) {
+    vars := mux.Vars(r)
+
+    instance := getInstanceAlias(vars["Instance"])
+    userId := vars["Id"]
+
+    beacon, err := getBeaconData(instance)
+
+    if err == nil {
+        delete(beacon.Users, userId)
+        err = updateBeaconField("Users", beacon.Users, instance)
+    }
+
+    writeResponse(err, w)
+}
+
+func handleUpdateBeaconAddress(w http.ResponseWriter, r *http.Request) {
+    vars := mux.Vars(r)
+
+    instance := getInstanceAlias(vars["Instance"])
+    address := vars["Address"]
+
+    err := updateBeaconField("BeaconAddress", address, instance)
+
+    writeResponse(err, w)
+}
+
+func handleUpdateBeaconToken(w http.ResponseWriter, r *http.Request) {
+    instance := getInstanceAlias(mux.Vars(r)["Instance"])
+
+    reqBody, err := ioutil.ReadAll(r.Body)
+    if err != nil {
+        writeResponse(err, w)
+        return
+    }
+
+    var token string
+
+    err = json.Unmarshal(reqBody, &token)
+    if err != nil {
+        writeResponse(err, w)
+        return
+    }
+
+    err = updateBeaconField("Token", token, instance)
+
+    writeResponse(err, w)
+}
+
+func handleCreate(r *http.Request) (int, error) {
+    reqBody, err := ioutil.ReadAll(r.Body)
+    if err != nil {
+        return http.StatusInternalServerError, err
+    }
+
+    var beaconInfo struct {
+        Address string
+        Token string
+        Users []string
+    }
+
+    err = json.Unmarshal(reqBody, &beaconInfo)
+    if err != nil {
+        return http.StatusInternalServerError, err
+    }
+
+    beacon := beaconData{"", beaconInfo.Address, beaconInfo.Token, make(userMap)}
+    for _, user := range beaconInfo.Users {
+        beacon.Users[user] = true
+    }
+
+    vmsTarget := fmt.Sprintf("http://%s/vms", beacon.BeaconAddress)
+
+    req, err := http.NewRequest("GET", vmsTarget, nil)
+    if err != nil {
+        return http.StatusInternalServerError, err
+    }
+
+    // Assuming user has permission to access token since they provided it
+    req.Header.Set(HEADER_TOKEN_KEY, beaconInfo.Token)
+
+    resp, err := http.DefaultClient.Do(req)
+    if err != nil {
+        return http.StatusInternalServerError, err
+    }
+
+    defer resp.Body.Close()
+
+    if resp.StatusCode != http.StatusOK {
+        return resp.StatusCode, errors.New("beacon error")
+    }
+
+    vmsBody, err := ioutil.ReadAll(resp.Body)
+    if err != nil {
+        return http.StatusInternalServerError, err
+    }
+
+    var vms []beaconStructs.VM
+
+    err = json.Unmarshal(vmsBody, &vms)
+    if err != nil {
+        return http.StatusInternalServerError, err
+    }
+
+    for _, vm := range vms {
+        beacon.InstanceAddress = fmt.Sprintf("%s:%s/%s", vm.Address, vm.Port, vm.Version)
+        err = addBeacon(beacon)
+    }
+
+    return http.StatusOK, nil
+}
